% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataRange.R
\name{DataRange}
\alias{DataRange}
\title{DataRange}
\usage{
DataRange(data, Col_name)
}
\arguments{
\item{data}{A data frame that contains the interested numeric data under certain column}

\item{Col_name}{A character in form of "column name"}
}
\value{
A vector contains two values, the minimum and maximum of the selected column
}
\description{
This function can compute the range of a column by using the column name from a dataset
}
